refactor torqued
